requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
watchdog>=4.0
loguru>=0.7
litellm>=1.40.0
langgraph>=0.2.0
chromadb>=0.4.0
tiktoken>=0.5.0
httpx>=0.27
fastapi>=0.110
uvicorn[standard]>=0.27
websockets>=12.0
python-multipart>=0.0.9
streamlit>=1.32
aiosqlite>=0.19
sqlalchemy>=2.0
bandit>=1.7
semgrep>=1.0
aiofiles>=23.0
cryptography>=42.0.0
jinja2>=3.1
python-dateutil>=2.8
psutil>=5.9.0